Implemented Camera Switching (#3)
Switching between a 3rd person perspective and top-down view implemented.

Note: If you adjust the code to try and blend between views the path the camera takes is erratic, so currently the switch is carried out instantly.

Also adjusted loather tracking distance as I think I set it a bit high, when I first did it. Though that's a problem for game design further down the line.